implemened the ORM (sqlalchemy) for the social login:
