+++ fapi/utils/lead_utils.py
@@ -2,11 +2,11 @@
 
 from sqlalchemy.orm import Session
 from fapi.db.database import SessionLocal
-from fapi.db.models import LeadCreate
+from fapi.db.schemas import LeadCreate
 from typing import Dict,Any
 from fastapi import HTTPException
 from sqlalchemy import func
-from fapi.db.schemas import LeadORM
+from fapi.db.models import LeadORM
 from datetime import datetime
 
 def fetch_all_leads_paginated(page: int, limit: int) -> Dict[str, any]:
